Update deprecated gbp.conf format.
As of version 0.6.0 of git-buildpackage [1] the git- and gbp- prefixes
were stripped from a number of configuration section names. This adds a
new template variable which changes the section name based on the
os-version (distro codename) and passes that back to the template.

I don't know if this is the right way to go about updating this and I
didn't add a comprehensive list of distributions that should use the
legacy section name. I opted to track distros using the legacy name
rather than distros using the static name since (in theory) the distros
using the legacy name are static and eventually won't be supported (when
we stop building any ROS distro that supports the older Debian and
Ubuntu releases).

Still to do:

- [ ] Verify the list of distributions supported that don't support the
  new config section name and add them to the distro list in
  `generate_debian`.

- [ ] Add tests for the new vs old header name for some sample distros.

- [ ] Add a timebomb or FIXUP comment so we can eventually remove this
  code when nothing on the build server relies on the old config
  version.

[1]: http://metadata.ftp-master.debian.org/changelogs/main/g/git-buildpackage/git-buildpackage_0.6.22_changelog